Refs #7270 Fix LENGTH_UNLIMITED macro double declaration

+++ src/cpp/fastrtps_deprecated/types/DynamicData.cpp
@@ -3440,7 +3440,7 @@ ReturnCode_t DynamicData::set_bool_value(
                 uint64_value_ = 0;
             }
         }
-        else if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) || id < type_->get_bounds())
+        else if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED || id < type_->get_bounds())
         {
             if (value)
             {
@@ -3513,7 +3513,7 @@ ReturnCode_t DynamicData::set_bool_value(
                 }
                 return ReturnCode_t::RETCODE_OK;
             }
-            else if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+            else if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                     id < type_->get_bounds())
             {
                 auto m_id = descriptors_.find(id);
@@ -4660,7 +4660,7 @@ ReturnCode_t DynamicData::insert_complex_value(
 {
     if (get_kind() == TK_SEQUENCE && type_->get_element_type()->equals(value->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4692,7 +4692,7 @@ ReturnCode_t DynamicData::insert_complex_value(
 {
     if (get_kind() == TK_SEQUENCE && type_->get_element_type()->equals(value->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4724,7 +4724,7 @@ ReturnCode_t DynamicData::insert_complex_value(
 {
     if (get_kind() == TK_SEQUENCE && type_->get_element_type()->equals(value->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4756,7 +4756,7 @@ ReturnCode_t DynamicData::insert_sequence_data(
     outId = MEMBER_ID_INVALID;
     if (get_kind() == TK_SEQUENCE)
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4825,7 +4825,7 @@ ReturnCode_t DynamicData::insert_map_data(
 {
     if (get_kind() == TK_MAP && type_->get_key_element_type()->equals(key->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4889,7 +4889,7 @@ ReturnCode_t DynamicData::insert_map_data(
     if (get_kind() == TK_MAP && type_->get_key_element_type()->equals(key->type_.get()) &&
             type_->get_element_type()->equals(value->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
@@ -4951,7 +4951,7 @@ ReturnCode_t DynamicData::insert_map_data(
     if (get_kind() == TK_MAP && type_->get_key_element_type()->equals(key->type_.get()) &&
             type_->get_element_type()->equals(value->type_.get()))
     {
-        if (type_->get_bounds() == static_cast<uint32_t>(dds::core::LENGTH_UNLIMITED) ||
+        if (type_->get_bounds() == dds::core::LENGTH_UNLIMITED ||
                 get_item_count() < type_->get_bounds())
         {
 #ifdef DYNAMIC_TYPES_CHECKING
